+++ CHANGELOG.md
@@ -5,6 +5,12 @@ CHANGELOG
 dev
 ---
 
+ ### New features
+
+ * ons-toolbar: Added new `static` attribute to avoid animations in the toolbar when pushing or popping pages.
+
+ ### Bug Fixes
+
  * ons-tabbar: Fixed [#2316](https://github.com/OnsenUI/OnsenUI/issues/2316).
  * ons-splitter-side: Fixed [#2271](https://github.com/OnsenUI/OnsenUI/issues/2271).
  * ons-navigator: Fixed [#2333](https://github.com/OnsenUI/OnsenUI/issues/2333).